requirements.deploy.txt
# AI Employee Vault - Deployment Dependencies (slim)
# Excludes heavy dev/browser packages for fast cloud builds

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Core
watchdog>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
aiohttp>=3.8.0
schedule>=1.2.0
tenacity>=8.0.0
psutil>=5.0.0

# Google/Gmail
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Social media
tweepy>=4.0.0
facebook-sdk>=3.0.0

# ERP
odoorpc>=0.9.0

# Git sync
gitpython>=3.0.0

# For platinum phase distributed operations
paramiko>=3.4.0
fabric>=3.2.0
cryptography>=41.0.0
PyGithub>=1.59.0
pydrive2>=1.19.0

# For monitoring
prometheus-client>=0.20.0
